Move createMockData to a helpers module
